+++ src/openfacefx/__init__.py
@@ -32,6 +32,8 @@ from .emotion import (EmotionEnvelope, bake_emotion, load_envelope,
 from .importers import (import_cues, detect_format, build_cue_track,
                        RHUBARB_TO_VISEME, PRESTON_BLAIR_TO_VISEME)
 from .importers_csv import read_csv
+from .inspect import (inspect_track, validate_asset, validate_file,
+                     detect_kind)
 from .export_unity import write_unity_anim, NAMING_PRESETS
 from .export_unreal_notifies import write_unreal_notifies, notifies_to_dict
 from .export_live2d import write_live2d_motion, lipsync_param_ids
@@ -85,6 +87,7 @@ __all__ = [
     "va_to_pose", "VA_TABLE", "VA_EMOTION_CHANNELS",
     "import_cues", "detect_format", "build_cue_track",
     "RHUBARB_TO_VISEME", "PRESTON_BLAIR_TO_VISEME", "read_csv",
+    "inspect_track", "validate_asset", "validate_file", "detect_kind",
     "write_unity_anim", "NAMING_PRESETS",
     "write_unreal_notifies", "notifies_to_dict",
     "write_live2d_motion", "lipsync_param_ids", "write_godot_anim",